Refactor all Brownian utilities.
1) Add caching for C++-based utilities.
2) Use typing for arguments.
3) Refactor benchmark.

This commit breaks the test for BrownianTree due to not being able to set w1.

+++ csrc/brownian_path.cpp
@@ -24,27 +24,27 @@ limitations under the License.
 
 #include "utils.hpp"
 
-BrownianPath::BrownianPath(float t0, torch::Tensor w0) {
-  cache.insert(std::pair<float, torch::Tensor>(t0, w0));
+BrownianPath::BrownianPath(double t0, torch::Tensor w0) {
+  cache.insert(std::pair<double, torch::Tensor>(t0, w0));
 }
 
-BrownianPath::BrownianPath(std::map<float, torch::Tensor> data) {
+BrownianPath::BrownianPath(std::map<double, torch::Tensor> data) {
   cache = data;
 }
 
-torch::Tensor BrownianPath::call(float t) {
+torch::Tensor BrownianPath::call(double t) {
   auto head = cache.begin();
   auto tail = cache.rbegin();
 
   if (t > tail->first) {
     auto w =
         tail->second + torch::randn_like(tail->second) * sqrt(t - tail->first);
-    cache.insert(std::pair<float, torch::Tensor>(t, w));
+    cache.insert(std::pair<double, torch::Tensor>(t, w));
     return w;
   } else if (t < head->first) {
     auto w =
         head->second + torch::randn_like(head->second) * sqrt(head->first - t);
-    cache.insert(std::pair<float, torch::Tensor>(t, w));
+    cache.insert(std::pair<double, torch::Tensor>(t, w));
     return w;
   } else if (cache.find(t) != cache.end()) {
     return cache.at(t);
@@ -52,27 +52,29 @@ torch::Tensor BrownianPath::call(float t) {
     auto lo = cache.lower_bound(t);
     auto hi = lo--;
     auto w = brownian_bridge(t, lo->first, hi->first, lo->second, hi->second);
-    cache.insert(std::pair<float, torch::Tensor>(t, w));
+    cache.insert(std::pair<double, torch::Tensor>(t, w));
     return w;
   }
 }
 
-void BrownianPath::insert(float t, torch::Tensor w) {
-  cache.insert(std::pair<float, torch::Tensor>(t, w));
+void BrownianPath::insert(double t, torch::Tensor w) {
+  cache.insert(std::pair<double, torch::Tensor>(t, w));
 }
 
 std::string BrownianPath::repr() const {
-  float t_head = cache.begin()->first;
-  float t_tail = cache.rbegin()->first;
+  double t_head = cache.begin()->first;
+  double t_tail = cache.rbegin()->first;
   return "BrownianPath(t0=" + format_float(t_head, 3) +
          ", t1=" + format_float(t_tail, 3) + ")";
 }
 
-std::map<float, torch::Tensor> BrownianPath::get_cache() const { return cache; }
+std::map<double, torch::Tensor> BrownianPath::get_cache() const {
+  return cache;
+}
 
-float BrownianPath::get_t_head() const { return cache.begin()->first; }
+double BrownianPath::get_t_head() const { return cache.begin()->first; }
 
-float BrownianPath::get_t_tail() const { return cache.rbegin()->first; }
+double BrownianPath::get_t_tail() const { return cache.rbegin()->first; }
 
 torch::Tensor BrownianPath::get_w_head() const { return cache.begin()->second; }
